•	. One of the biggest threats can be the interception of the traffic when the mobile payment is in process. This can lead to identity theft
We are using a method widely known as the Secure Electronic Transaction
to solve the security issues of mobile payment. The main purpose of the
project it to solve issues like interception of payment info of the user
and using such information for malicious purposes. SET achieves this by
using the concept of Dual Signature. With SET, the user is given a
digital certificate and a transaction is conducted and verified using a
combination of digital signature and digital certificates among the
user, the merchant and the user’s bank.